+++ documentation/changelog.md
@@ -39,6 +39,7 @@ description: List of all the changes and current versions.
 + Add Custom Vehicle Rules for engine, unlock, lock, doors, enter, and exit
 + Add Custom Storage Rules for opening
 + Add Custom Interior Rules for enter, exit, unlock, and lock
++ Add / Remove Faction(s) from Interiors with Function
 ```
 
 ## 2.0.3